+++ docs/source/sphinx_gallery/geometry/plot_ellipse.py
@@ -37,5 +37,5 @@ def breakdown():
 
 
 if __name__ == '__main__':
-    plt.rc('savefig', dpi=300, bbox='tight', pad_inches=0)
+    plt.rc('savefig', dpi=300, bbox_inches='tight', pad_inches=0)
     main()